Refactor serverless functions, added CI/CD to them as well
