+++ serverless/bridge-tasks/.funcignore
