+++ serverless/bridge-tasks/PriceApiFetch/function.json
