+++ serverless/bridge-tasks/ScrtSender/function.json
