+++ serverless/bridge-tasks/SignerHealth/function.json
