+++ serverless/bridge-tasks/TotalLockedFetch/function.json
