+++ serverless/bridge-tasks/TotalLockedFetch/index.ts
