+++ serverless/bridge-tasks/package.json
@@ -10,13 +10,15 @@
     "test": "echo \"No tests yet...\""
   },
   "author": "",
-  "dependencies": {
+  "devDependencies": {
     "@azure/functions": "^1.0.1-beta1",
-    "mongodb": "^3.6.3",
-    "node-fetch": "^2.3.0",
     "npm-run-all": "^4.1.5",
+    "typescript": "^3.3.3"
+  },
+  "dependencies": {
+    "mongodb": "^3.6.6",
+    "node-fetch": "^2.6.1",
     "secretjs": "^0.11.0",
-    "typescript": "^3.3.3",
-    "web3": "^1.3.0"
+    "web3": "^1.3.5"
   }
 }
