+++ serverless/secretswap-tasks/.funcignore
